scripts/requirements.txt
easy-equities-client>=0.1.0
python-dotenv>=1.0.0
